Counter fixes
